Store original JS file for *Grimoire
This is the original JS file from VizGrimoireJS, which is pulled in by
VizGrimoireR (see 09f142e for a pointer to that repo).  VizGrimoireJS
keeps it minified, but this is the unminified version.  I'm committing
the original here so that I can keep track of the changes I make while
creating custom charts.