Note chart display options, change time-to-fix
Make note of the different display options for charts all in one place.
It may be that I should be passing these options from the python script
in VizGrimoireR, but I'm not.  Change the default chart colors (mostly
to see where that change should be made).  Change labels for the
time-to-fix chart.  Optionally show only some of the lines on the
time-to-fix chart.