Update closed chart with different color
Change the color of the bars in the "bugs-closed" chart to match the
"closed" colors in the bugs-opened-v-closed chart, for consistency.  I
think it's easier to see the relationship between these three charts if
opened is always blue and closed is always green.

Note: to do this, I changed the order of colors in vizgrimoire.js (see
42dff944 for an example of this) so that green is used first.  That
makes a chart with only one set of data (in this case, number of bugs
closed) show up in green. See commit 5c934ab in the geonode-orgs branch
of VizGrimoireR
(https://github.com/VizGrimoire/VizGrimoireR/commit/5c934abca98f4402b9d66e3f641531787c4b36bf)
for a full view of how I changed the dashboard to view these charts
separately.  That index.html file is copied to the dashboard directory
by the vg-github.py script.