Make a full image of Malawi catalog page
Add page top (with satellite image) and middle region (with search
bar) to the Malawi catalog screenshot of commit 3113000f.  Because my
browser window was too short to hold them all together, this is
actually a composition of several successively-scrolled screenshots.
Please note the painstakingly corrected tooltip portion, with manual
re-shading to get the middle-to-bottom transition undershadow right.